+++ templates/bidirectional-widget-plugin/{{ cookiecutter.project_name }}/setup.cfg
@@ -24,4 +24,4 @@ where=src
 
 [options.entry_points]
 deephaven.plugin =
-    registration_cls = {{ cookiecutter.__py_namespace }}._register:{{ cookiecutter.__registration_name }}
+    registration_cls = {{ cookiecutter.__py_namespace }}.register:{{ cookiecutter.__registration_name }}
