+++ templates/bidirectional-widget-plugin/{{ cookiecutter.project_name }}/src/{{ cookiecutter.__src_folder_name }}/js_plugin.py
